+++ src/UtilEffects/RateRandomizer/RateRandomizerPS.cpp
@@ -1,6 +1,6 @@
 #include "RateRandomizerPS.h"
 
-RateRandomizerPS::RateRandomizerPS(uint16_t BaseRate, int16_t RateRangeMin, int16_t RateRangeMax, uint16_t Rate)
+RateRandomizerPS::RateRandomizerPS(uint16_t BaseRate, int32_t RateRangeMin, int32_t RateRangeMax, uint16_t Rate)
     : rateRangeMin(RateRangeMin), rateRangeMax(RateRangeMax)  //
 {
     //do the pointer binding for the base rate so it can be externally tied
@@ -11,12 +11,14 @@ RateRandomizerPS::RateRandomizerPS(uint16_t BaseRate, int16_t RateRangeMin, int1
     bindClassRatesPS();
 }
 
+//Resets the outputRate to the baseRate. 
+//Also sets "active" to false, turning off the utility.
 void RateRandomizerPS::resetToBaseRate() {
     outputRate = *baseRate;
     active = false;
 }
 
-//update call is pretty simple
+//The update call is pretty simple
 //we take the base rate and add on a random value between the rate ranges
 //this is stored in tempOut, a int32, so we can check for int16 over/under flow
 void RateRandomizerPS::update() {
